Crate pallet_transaction_multi_payment

Re-exports

pub use pallet::*;

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

weights

Autogenerated weights for pallet_transaction_multi_payment

Structs

AddTxAssetOnAccount

Type to automatically add a fee currency for an account on account creation.

CurrencyBalanceCheck

Signed extension that checks for the set_currency call and in that case, it checks the account balance

DepositAll

Deposits all fees to some account

RemoveTxAssetOnKilled

Type to automatically remove the fee currency for an account on account deletion.

TransferFees

Implements the transaction payment for native as well as non-native currencies

WithdrawFees

Implements the transaction payment for native as well as non-native currencies

Enums

PaymentInfo

PaymentWithdrawResult

Traits

CurrencyWithdraw

DepositFee

Handler for dealing with fees

TransactionMultiPaymentDataProvider

Helper method for providing some data that are needed in OnChargeTransaction

Functions

error_to_invalid

convert an Error to a custom InvalidTransaction with the inner code being the error number.

Type Definitions

Price

Spot price type