Trait pallet_transaction_multi_payment::TransactionMultiPaymentDataProvider

pub trait TransactionMultiPaymentDataProvider<AccountId, AssetId, Price> {
    fn get_currency_and_price(
        who: &AccountId
    ) -> Result<(AssetId, Option<Price>), DispatchError>;

    fn get_fee_receiver() -> AccountId;
}

Helper method for providing some data that are needed in OnChargeTransaction

Required methods

fn get_currency_and_price(
who: &AccountId
) -> Result<(AssetId, Option<Price>), DispatchError>

Get a fee currency set by an account and its price

fn get_fee_receiver() -> AccountId

Returns the account where fees are deposited