Trait pallet_transaction_multi_payment::pallet::Config

pub trait Config: Config + Config {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type Currency: Currency<Self::AccountId> + Send + Sync;
    type MultiCurrency: MultiCurrency<Self::AccountId> + MultiCurrencyExtended<Self::AccountId, CurrencyId = AssetId, Balance = Balance, Amount = Amount>;
    type AMMPool: AMM<Self::AccountId, AssetId, AssetPair, Balance>;
    type WeightInfo: WeightInfo;
    type WithdrawFeeForSetCurrency: Get<Pays>;
    type WeightToFee: WeightToFeePolynomial<Balance = Balance>;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

Because this pallet emits events, it depends on the runtime’s definition of an event.

type Currency: Currency<Self::AccountId> + Send + Sync

The currency type in which fees will be paid.

type MultiCurrency: MultiCurrency<Self::AccountId> + MultiCurrencyExtended<Self::AccountId, CurrencyId = AssetId, Balance = Balance, Amount = Amount>

Multi Currency

type AMMPool: AMM<Self::AccountId, AssetId, AssetPair, Balance>

AMM pool to swap for native currency

type WeightInfo: WeightInfo

Weight information for the extrinsics.

type WithdrawFeeForSetCurrency: Get<Pays>

Should fee be paid for setting a currency

type WeightToFee: WeightToFeePolynomial<Balance = Balance>

Convert a weight value into a deductible fee based on the currency type.