Trait pallet_transaction_multi_payment::pallet::Config

pub trait Config: Config + Config {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type AcceptedCurrencyOrigin: EnsureOrigin<Self::Origin>;
    type Currencies: MultiCurrencyExtended<Self::AccountId>;
    type SpotPriceProvider: SpotPriceProvider<<<Self as Config>::Currencies as MultiCurrency<<Self as Config>::AccountId>>::CurrencyId, Price = Price>;
    type WeightInfo: WeightInfo;
    type WithdrawFeeForSetCurrency: Get<Pays>;
    type WeightToFee: WeightToFeePolynomial<Balance = <<Self as Config>::Currencies as MultiCurrency<<Self as Config>::AccountId>>::Balance>;
    type NativeAssetId: Get<<<Self as Config>::Currencies as MultiCurrency<<Self as Config>::AccountId>>::CurrencyId>;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

Because this pallet emits events, it depends on the runtime’s definition of an event.

type AcceptedCurrencyOrigin: EnsureOrigin<Self::Origin>

The origin which can add/remove accepted currencies

type Currencies: MultiCurrencyExtended<Self::AccountId>

Multi Currency

type SpotPriceProvider: SpotPriceProvider<<<Self as Config>::Currencies as MultiCurrency<<Self as Config>::AccountId>>::CurrencyId, Price = Price>

Spot price provider

type WeightInfo: WeightInfo

Weight information for the extrinsics.

type WithdrawFeeForSetCurrency: Get<Pays>

Should fee be paid for setting a currency

type WeightToFee: WeightToFeePolynomial<Balance = <<Self as Config>::Currencies as MultiCurrency<<Self as Config>::AccountId>>::Balance>

Convert a weight value into a deductible fee based on the currency type.

type NativeAssetId: Get<<<Self as Config>::Currencies as MultiCurrency<<Self as Config>::AccountId>>::CurrencyId>

Native Asset